pyproject.toml
[project]
name = "python_queue_benchmark"
version = "0.1.0"
description = "Benchmark of Python Task Queue Libraries"
authors = [
    {name = "Steven Yue", email = "hey.steveny@gmail.com"},
]
dependencies = [
    "rq>=2.7.0",
    "dramatiq[redis,watch]>=2.1.0",
    "huey==2.5.1",
    "arq==0.26.3",
    "supervisor==4.2.5",
    "celery[redis]==5.4.0",
    "taskiq==0.11.19",
    "taskiq-redis==1.0.9",
    "matplotlib>=3.10.7",
    "procrastinate>=3.5.3",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}
